Make /transactions in the API more typed
Previously it was taking any bytes and sending them as is. This means
that the sender had to bother with encoding their transactions in bytes
themselves.

Now it takes a list of signed transactions and handles the encoding.

The API now provides `_transactionsRaw` which has the old type.

`halgo node send` no longer allows sending arbitrary bytes, it will
always try to parse its input as signed transactions. Feel free to
implement something like a `--raw` flag, if you need this.