Implement getBlock helper function
Problem:
Retrieving the block for round for which no block exists results into the error.

Solution:
Introduce a `getBlock` function which catches error that has block absence for the round as the cause and return Nothing in this case.
Other errors would be rethrown.

Note:
This commit is rework of PR #1.