Adapt API to changes made by RandLabs
Problem:
RandLabs updated API of Algorand node and indexer. For example, we know
that `/health` and `/versions` endpoints apparently won't exist anymore.
Also they told that in future node will be used only for transactions.

Solution:
- Split Node and Indexer API types
- Add deprecation warning for some node commands and helpers
- Add few missing fields to indexer types and delete committed fields from node one
- Rename types for API and delete `Api` type
- Delete `/health` and `/versions` endpoints

Note: for now there are no check that we operate at expected network,
because we can not check it any more.
In `connectToNode` and `connectToIndexer` will return passed network
without any check. We will make some changes to this when we receive
an answer from RandLabs.