Remove ApiAny and network check
Problem: ApiAny has `health` and `versions` endpoints. First of all,
none of them are present in the new indexer. The former is said to be
internal and thus shouldn't be available in SDK. The latter is being
discussed. Maybe it will be added to the indexer, maybe it will be
removed from the node.

Solution: now there are no shared endpoints, so ApiAny is useless and
we remove it. Even if `versions` is restored, having ApiAny for just
1 endpoint doesn't seem very useful. For now we remove `versions`
completely until we get clarifications from RandLabs.
`/health` is gone completely since it's internal.