Uncomment asCreator field
Problem: `asCreator` field is commented out because the old indexer
always returns `""` for this field (and we can't parse it because it's
not a valid address).
However, this bug is fixed in the new indexer and there is no reason
to use the old one anymore.

Solution: uncomment this field.