Add `/blocks` endpoint to the indexer API
Problem: we have an endpoint to query block data in node API,
but according to RandLabs this way of getting data will be
deprecated soon. It's recommended to get data from the indexer
instead.

Solution: there is `/blocks` endpoint in the indexer as well,
so we add it to IndexerApi.
The corresponding command got added to `halgo`.
Getting block via node API is deprecated.
We add new TransactionResp and BlockResp types because transactions
returned by the indexer structurally differ from transactions in
node API.