[SDAO-497] Make `OnComplete` as sum types
Problem: The fix to `OnComplete` parsers for indexer affects the encoding of it as
well which result in the encoding changes in the test. This is not
expected and the encoding shouldn't change.

Solution: Make `OnComplete` as sum types, its `AlgoMessagePack` instance
still expects it to be a number while its JSON instance expect it to be a string
instead, fix the tests.