[#16]: Clean up tests
Problem:
Currently we have a lot of generators and utils and test files, some of
them exports that staff. It's hard to know what we have, which leads to
code duplication and dirty structure.

Solution:
Provide `Gen` module with generators for tests, `Domain` module with some
mock data and instances and move utils to `Utils` module.