Mention that `isValidGroup` is not ready
Problem: in Staker bridge repository we rely on `isValidGroup` function,
but it is too strict. It tries to compute the transaction group and
compare it to what is returned by the indexer.

The issue is that, we do not support all types of transactions.
We do not parse properly some transaction types, but then try to
serialize them when computing tx group, and that simply does not work.

Solution: leave a comment in module with `isValidGroup` that it is not
guaranteed to work.

And add a test in this function that proves that it does not work in
many situations.

Frankly speaking, the point of this check is vague for me, we suggest
that our SDK (that supports only part of the API) will be able to
correctly find issues with the indexer.
But I'd believe the indexer more than our SDK in 100% of the cases,
really. Maybe in 99%, but only when we provide the full support.
At this point, having these checks seems just meaningless. But I'm
leaving that functionality because some research has been made to write
it, and maybe one day we manage to complete our SDK.